Parse Args, moving on to other tests
