+++ src/roll/generate-result/apply-modifiers.ts
