+++ src/roll/generate-result/index.ts
