+++ tests/generate-result.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test } from 'bun:test'
 
-import { dicePoolFactory, StandardDie } from '../src/Die'
-import generateResult from '../src/roll/generate-results'
-import { InvalidUniqueError } from '../src/roll/generate-results/apply-modifiers'
+import { StandardDie } from '../src/Die'
+import generateResult from '../src/roll/generate-result'
+import { InvalidUniqueError } from '../src/roll/generate-result/apply-modifiers'
 import { Modifiers } from '../src/types/options'
 import { RollParameters } from '../src/types/parameters'
